Switch to use SLF4J via Logback in Magenta-CLI
Decided to use SLF4J in the magenta-lib project so that logback isn't a requirement.
But magenta-cli abuses logbacks logger.setlevel stuff for config, so it assumes it's logback, which is I think safe.
It now logs to stdout (slightly more messily), and to syslog on user2.
Where it logs should be configurable at runtime if at all possible, but that's a later story.